# artifacts/api-server/src/routes/nutterx/admin.ts
import { Router, Response } from "express";
import { User } from "../../models/User";
import { ServiceRequest } from "../../models/ServiceRequest";
import { Chat } from "../../models/Chat";
import { authenticate, requireAdmin, AuthRequest } from "../../middlewares/auth";
import { formatRequest } from "./requests";

const router = Router();

router.get("/users", authenticate, requireAdmin, async (_req: AuthRequest, res: Response): Promise<void> => {
  try {
    const users = await User.find().select("-password").sort({ createdAt: -1 });
    res.json(users);
  } catch {
    res.status(500).json({ message: "Failed to fetch users" });
  }
});

router.get("/requests", authenticate, requireAdmin, async (_req: AuthRequest, res: Response): Promise<void> => {
  try {
    const requests = await ServiceRequest.find()
      .populate("user", "name email")
      .sort({ createdAt: -1 });
    res.json(requests.map(formatRequest));
  } catch {
    res.status(500).json({ message: "Failed to fetch requests" });
  }
});

router.put("/requests/:id", authenticate, requireAdmin, async (req: AuthRequest, res: Response): Promise<void> => {
  try {
    const { status, adminNotes } = req.body;

    const update: Record<string, unknown> = { status, adminNotes };

    if (status === "completed") {
      update["completedAt"] = new Date();
      update["subscriptionEndsAt"] = new Date(Date.now() + 30 * 24 * 60 * 60 * 1000);
    }

    const request = await ServiceRequest.findByIdAndUpdate(
      req.params["id"],
      update,
      { new: true }
    ).populate("user", "name email");

    if (!request) {
      res.status(404).json({ message: "Request not found" });
      return;
    }

    res.json(formatRequest(request));
  } catch {
    res.status(500).json({ message: "Failed to update request" });
  }
});

router.get("/subscriptions", authenticate, requireAdmin, async (_req: AuthRequest, res: Response): Promise<void> => {
  try {
    const now = new Date();
    const requests = await ServiceRequest.find({
      status: "completed",
      subscriptionEndsAt: { $gt: now },
    })
      .populate("user", "name email")
      .sort({ subscriptionEndsAt: 1 });

    const subscriptions = requests.map((r) => {
      const msLeft = r.subscriptionEndsAt!.getTime() - now.getTime();
      const daysRemaining = Math.max(0, Math.ceil(msLeft / (1000 * 60 * 60 * 24)));
      return {
        _id: r._id,
        user: r.user,
        serviceName: r.serviceName,
        completedAt: r.completedAt,
        subscriptionEndsAt: r.subscriptionEndsAt,
        daysRemaining,
      };
    });

    res.json(subscriptions);
  } catch {
    res.status(500).json({ message: "Failed to fetch subscriptions" });
  }
});

router.get("/chats", authenticate, requireAdmin, async (_req: AuthRequest, res: Response): Promise<void> => {
  try {
    const chats = await Chat.find()
      .populate("participants", "name email avatar role")
      .sort({ updatedAt: -1 });
    res.json(chats);
  } catch {
    res.status(500).json({ message: "Failed to fetch chats" });
  }
});

export default router;