# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nutterx Technologies API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: services
    description: Service catalog operations
  - name: requests
    description: Service request operations
  - name: chats
    description: Chat and messaging operations
  - name: admin
    description: Admin operations
  - name: users
    description: User operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /services:
    get:
      operationId: getServices
      tags: [services]
      summary: Get all available services (public)
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a single service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /requests:
    post:
      operationId: createRequest
      tags: [requests]
      summary: Submit a service request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestBody"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
    get:
      operationId: getMyRequests
      tags: [requests]
      summary: Get current user's service requests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User's requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRequest"

  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a single service request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"

  /chats:
    get:
      operationId: getChats
      tags: [chats]
      summary: Get all chats for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"

  /chats/group:
    post:
      operationId: createGroupChat
      tags: [chats]
      summary: Create a group chat (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Group chat created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

  /chats/{chatId}/messages:
    get:
      operationId: getChatMessages
      tags: [chats]
      summary: Get messages for a chat
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [chats]
      summary: Send a message in a chat
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /chats/direct/{userId}:
    post:
      operationId: startDirectChat
      tags: [chats]
      summary: Start or get a direct chat with a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Direct chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Get all users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /admin/requests:
    get:
      operationId: adminGetRequests
      tags: [admin]
      summary: Get all service requests (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRequest"

  /admin/requests/{id}:
    put:
      operationId: adminUpdateRequest
      tags: [admin]
      summary: Update a service request status (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequestStatus"
      responses:
        "200":
          description: Request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"

  /admin/subscriptions:
    get:
      operationId: adminGetSubscriptions
      tags: [admin]
      summary: Get all active subscriptions (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: string
      required:
        - message

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - name
        - email
        - password

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        avatar:
          type: string
      required:
        - _id
        - name
        - email
        - role

    Service:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        features:
          type: array
          items:
            type: string
        icon:
          type: string
        category:
          type: string
        popular:
          type: boolean
      required:
        - _id
        - title
        - description

    CreateServiceRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        features:
          type: array
          items:
            type: string
        icon:
          type: string
        category:
          type: string
        popular:
          type: boolean
      required:
        - title
        - description

    CreateRequestBody:
      type: object
      properties:
        serviceId:
          type: string
        serviceName:
          type: string
        description:
          type: string
        requirements:
          type: string
      required:
        - serviceName
        - description

    ServiceRequest:
      type: object
      properties:
        _id:
          type: string
        user:
          $ref: "#/components/schemas/User"
        serviceId:
          type: string
        serviceName:
          type: string
        description:
          type: string
        requirements:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        completedAt:
          type: string
          format: date-time
        subscriptionEndsAt:
          type: string
          format: date-time
        daysRemaining:
          type: integer
        adminNotes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - _id
        - serviceName
        - description
        - status

    UpdateRequestStatus:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        adminNotes:
          type: string
      required:
        - status

    Chat:
      type: object
      properties:
        _id:
          type: string
        type:
          type: string
          enum: [direct, group]
        name:
          type: string
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        lastMessage:
          $ref: "#/components/schemas/Message"
        unreadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - _id
        - type
        - participants

    Message:
      type: object
      properties:
        _id:
          type: string
        chatId:
          type: string
        sender:
          $ref: "#/components/schemas/User"
        content:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - _id
        - chatId
        - content

    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    CreateGroupRequest:
      type: object
      properties:
        name:
          type: string
        participantIds:
          type: array
          items:
            type: string
      required:
        - name
        - participantIds

    Subscription:
      type: object
      properties:
        _id:
          type: string
        user:
          $ref: "#/components/schemas/User"
        serviceName:
          type: string
        completedAt:
          type: string
          format: date-time
        subscriptionEndsAt:
          type: string
          format: date-time
        daysRemaining:
          type: integer
      required:
        - _id
        - serviceName